changed: Break out non-async specific interrupt functionality out of async_hal module
